requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.3.4
anyio==3.7.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distro==1.8.0
dnspython==2.3.0
elasticsearch==8.8.2
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
loguru==0.7.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.24.3
outcome==1.2.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.11
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.6.3
urllib3>=1.21.1,<3
wcwidth==0.2.6
wikipedia==1.4.0
yarl==1.9.2